+++ R/autoplot-cps-spct.r
@@ -24,6 +24,9 @@
 #'   as arguments. A list with \code{numeric} and/or \code{character} values is
 #'   also accepted.
 #' @param annotations a character vector
+#' @param geom character The name of a ggplot geometry, currently only
+#'   \code{"area"}, \code{"spct"} and \code{"line"}. The default \code{NULL}
+#'   selects between them based on \code{stacked}.
 #' @param text.size numeric size of text in the plot decorations.
 #' @param idfactor character Name of an index column in data holding a
 #'   \code{factor} with each spectrum in a long-form multispectrum object
@@ -49,6 +52,7 @@ cps_plot <- function(spct,
                      span,
                      wls.target,
                      annotations,
+                     geom,
                      text.size,
                      idfactor,
                      facets,
@@ -58,6 +62,10 @@ cps_plot <- function(spct,
                      ...) {
   if (!is.cps_spct(spct)) {
     stop("cps_plot() can only plot cps_spct objects.")
+  }
+  if (!is.null(geom) && !geom %in% c("area", "line", "spct")) {
+    warning("'geom = ", geom, "' not supported, using default instead.")
+    geom <- NULL
   }
   if (is.null(ylim) || !is.numeric(ylim)) {
     ylim <- rep(NA_real_, 2L)
@@ -109,11 +117,11 @@ cps_plot <- function(spct,
     # remove cps_spct class before melting as it invalidates expectations
     rmDerivedSpct(spct)
     spct <- tidyr::pivot_longer(data = spct,
-                                cols = tidyselect::all_of(cps.cols),
+                                cols = tidyr::all_of(cps.cols),
                                 names_to = "scan",
                                 values_to = "cps")
     setCpsSpct(spct, multiple.wl = NULL) # guessed from data
-    plot <- ggplot(spct, aes_(x = ~w.length, y = ~cps, linetype = ~scan))
+    plot <- ggplot(spct, aes(x = .data[["w.length"]], y = .data[["cps"]], linetype = .data[["scan"]]))
     temp <- find_idfactor(spct = spct,
                           idfactor = idfactor,
                           facets = facets,
@@ -122,7 +130,7 @@ cps_plot <- function(spct,
     plot <- plot + temp$ggplot_comp
     annotations <- temp$annotations
   } else {
-    plot <- ggplot(spct, aes_(x = ~w.length, y = ~cps))
+    plot <- ggplot(spct, aes(x = .data[["w.length"]], y = .data[["cps"]]))
     temp <- find_idfactor(spct = spct,
                           idfactor = idfactor,
                           facets = facets,
@@ -147,6 +155,9 @@ cps_plot <- function(spct,
     }
   }
 
+  if (!is.null(geom) && geom %in% c("area", "spct")) {
+    plot <- plot + geom_spct(fill = "black", colour = NA, alpha = 0.2)
+  }
   plot <- plot + geom_line(na.rm = na.rm)
   plot <- plot + labs(x = "Wavelength (nm)", y = s.cps.label)
 
@@ -230,6 +241,9 @@ cps_plot <- function(spct,
 #'   also accepted.
 #' @param annotations a character vector ("summaries" is ignored). For details
 #'   please see sections Plot Annotations and Title Annotations.
+#' @param geom character The name of a ggplot geometry, currently only
+#'   \code{"area"}, \code{"spct"} and \code{"line"}. The default \code{NULL}
+#'   selects between them based on \code{stacked}.
 #' @param time.format character Format as accepted by \code{\link[base]{strptime}}.
 #' @param tz character Time zone to use for title and/or subtitle.
 #' @param norm numeric normalization wavelength (nm) or character string
@@ -261,16 +275,14 @@ cps_plot <- function(spct,
 #' @examples
 #'
 #' autoplot(white_led.cps_spct)
+#' autoplot(white_led.cps_spct, geom = "spct")
 #' autoplot(white_led.cps_spct, norm = "max")
-#' autoplot(white_led.cps_spct, annotations = "")
 #'
 #' two_leds.mspct <-
 #'   cps_mspct(list("LED 1" = white_led.cps_spct,
 #'                  "LED 2" = white_led.cps_spct / 2))
 #' autoplot(two_leds.mspct)
 #' autoplot(two_leds.mspct, idfactor = "Spectra")
-#' autoplot(two_leds.mspct, facets = 1) # one column
-#' autoplot(two_leds.mspct, facets = 2) # two columns
 #' autoplot(two_leds.mspct, plot.data = "mean")
 #'
 #' @family autoplot methods
@@ -288,6 +300,7 @@ autoplot.cps_spct <-
            span = NULL,
            wls.target = "HM",
            annotations = NULL,
+           geom = "line",
            time.format = "",
            tz = "UTC",
            text.size = 2.5,
@@ -332,6 +345,7 @@ autoplot.cps_spct <-
              span = span,
              wls.target = wls.target,
              annotations = annotations,
+             geom = geom,
              text.size = text.size,
              idfactor = idfactor,
              facets = facets,