+++ R/autoplot-raw-spct.r
@@ -26,6 +26,9 @@
 #'   as arguments. A list with \code{numeric} and/or \code{character} values is
 #'   also accepted.
 #' @param annotations a character vector.
+#' @param geom character The name of a ggplot geometry, currently only
+#'   \code{"area"}, \code{"spct"} and \code{"line"}. The default \code{NULL}
+#'   selects between them based on \code{stacked}.
 #' @param text.size numeric size of text in the plot decorations.
 #' @param idfactor character Name of an index column in data holding a
 #'   \code{factor} with each spectrum in a long-form multispectrum object
@@ -51,6 +54,7 @@ raw_plot <- function(spct,
                      span,
                      wls.target,
                      annotations,
+                     geom,
                      relative,
                      text.size,
                      idfactor,
@@ -60,6 +64,10 @@ raw_plot <- function(spct,
                      ...) {
   if (!is.raw_spct(spct)) {
     stop("raw_plot() can only plot response_spct objects.")
+  }
+  if (!is.null(geom) && !geom %in% c("area", "line", "spct")) {
+    warning("'geom = ", geom, "' not supported, using default instead.")
+    geom <- NULL
   }
   if (is.null(ylim) || !is.numeric(ylim)) {
     ylim <- rep(NA_real_, 2L)
@@ -123,11 +131,11 @@ raw_plot <- function(spct,
     # remove raw_spct class before melting as it invalidates expectations
     rmDerivedSpct(spct)
     spct <- tidyr::pivot_longer(data = spct,
-                                cols = tidyselect::all_of(counts.cols),
+                                cols = tidyr::all_of(counts.cols),
                                 names_to = "scan",
                                 values_to = "counts")
     setRawSpct(spct, multiple.wl = NULL) # guessed from data
-    plot <- ggplot(spct) + aes_(x = ~w.length, y = ~counts, linetype = ~scan)
+    plot <- ggplot(spct) + aes(x = .data[["w.length"]], y = .data[["counts"]], linetype = .data[["scan"]])
     temp <- find_idfactor(spct = spct,
                           idfactor = idfactor,
                           facets = facets,
@@ -136,7 +144,7 @@ raw_plot <- function(spct,
     plot <- plot + temp$ggplot_comp
     annotations <- temp$annotations
   } else {
-    plot <- ggplot(spct) + aes_(x = ~w.length, y = ~counts)
+    plot <- ggplot(spct) + aes(x = .data[["w.length"]], y = .data[["counts"]])
     temp <- find_idfactor(spct = spct,
                           idfactor = idfactor,
                           facets = facets,
@@ -174,6 +182,9 @@ raw_plot <- function(spct,
     }
   }
 
+  if (!is.null(geom) && geom %in% c("area", "spct")) {
+    plot <- plot + geom_spct(fill = "black", colour = NA, alpha = 0.2)
+  }
   plot <- plot + geom_line(na.rm = na.rm)
   plot <- plot + labs(x = "Wavelength (nm)", y = s.counts.label)
 
@@ -260,6 +271,9 @@ raw_plot <- function(spct,
 #'   also accepted.
 #' @param annotations a character vector ("summaries" is ignored). For details
 #'   please see sections Plot Annotations and Title Annotations.
+#' @param geom character The name of a ggplot geometry, currently only
+#'   \code{"area"}, \code{"spct"} and \code{"line"}. The default \code{NULL}
+#'   selects between them based on \code{stacked}.
 #' @param time.format character Format as accepted by \code{\link[base]{strptime}}.
 #' @param tz character Time zone to use for title and/or subtitle.
 #' @param norm numeric normalization wavelength (nm) or character string "max"
@@ -289,16 +303,18 @@ raw_plot <- function(spct,
 #'
 #' @examples
 #'
-#' autoplot(white_led.raw_spct)
-#' autoplot(white_led.raw_spct, annotations = "")
+#' low_res.raw_spct <- thin_wl(white_led.raw_spct,
+#'                             max.wl.step = 20,
+#'                             max.slope.delta = 0.05,
+#'                             col.names = "counts_3")
+#' autoplot(low_res.raw_spct)
+#' autoplot(low_res.raw_spct, annotations = "")
 #'
 #' two_leds.mspct <-
-#'   raw_mspct(list("LED 1" = white_led.raw_spct,
-#'                  "LED 2" = white_led.raw_spct))
+#'   raw_mspct(list("LED 1" = low_res.raw_spct,
+#'                  "LED 2" = low_res.raw_spct))
 #' autoplot(two_leds.mspct)
-#' autoplot(two_leds.mspct, idfactor = "Spectra")
 #' autoplot(two_leds.mspct, facets = 1) # one column
-#' autoplot(two_leds.mspct, facets = 2) # two columns
 #'
 #' @family autoplot methods
 #'
@@ -313,6 +329,7 @@ autoplot.raw_spct <-
            span = NULL,
            wls.target = "HM",
            annotations = NULL,
+           geom = "line",
            time.format = "",
            tz = "UTC",
            norm = "skip",
@@ -358,6 +375,7 @@ autoplot.raw_spct <-
              wls.target = wls.target,
              pc.out = pc.out,
              annotations = annotations,
+             geom = geom,
              norm = norm,
              text.size = text.size,
              idfactor = idfactor,