+++ R/color-chart.R
@@ -74,16 +74,19 @@ color_chart <- function(colors = grDevices::colors(),
                                    width = 3))
                )
   # we build the plot
-  p <- ggplot(colors.df, aes_(~x, ~y, fill = ~color))
+  p <- ggplot(colors.df,
+              ggplot2::aes(x = .data[["x"]],
+                           y = .data[["y"]],
+                           fill = .data[["color"]]))
   if (use.names) {
-    p <- p + aes_(label = ~color.names)
+    p <- p + ggplot2::aes(label = .data[["color.names"]])
   } else {
-    p <- p + aes_(label = ~idx)
+    p <- p + ggplot2::aes(label = .data[["idx"]])
   }
   p <- p +
     geom_tile(color = grid.color) +
     scale_fill_identity() +
-    geom_text(size = text.size, aes_(color = ~text.color)) +
+    geom_text(size = text.size, ggplot2::aes(color = .data[["text.color"]])) +
     scale_color_identity()
   p + theme_void()
 }